Analyze a grayscale keyframe to split it into cells, and output the median values from the pixels in each cell to a CSV file.
The constants VerticalCellCount and HorizontalCellCount declare how many cells
to split each frame into. FrameAnalysisCSVFile specifies the output filename.

analyzeGrayscaleFrame() splits a frame into a grid of cells. Cell dimensions
will differ between cells if the video's native resolution does not evenly
divide by the cell count, but the method attempts to split the cells as evenly
as possible. It calls cellMedian() to retrieve the median value for each cell,
then passes those values to outputFrameAnalysis().

outputFrameAnalysis() determines the video timestamp from the frame's
best_effort_timestamp and the stream's time base. It outputs the timestamp and
the cell median values to the CSV file according to the specifications given in
the challenge <https://gist.github.com/ehmo/ece1b13a101b416c6b3c928ad11e1776>.

Note that it isn't clear from the challenge what the bit depth of the grayscale
image should be, but it's assumed to be 8bpp. Also note that the cell ordering
isn't defined in the challenge, but it's assumed to be row-major order.

cellMedian() copies all of the pixel data for the given cell location and
dimensions to a vector, then calls median() to retrieve the median value.

median() uses the nth_element algorithm to partially sort the input vector and
retrieve the median value. If there are an odd number of items in the set, this
is trivial, but if the set contains an even number of items, the true median
is the average of the two middle items. Calling nth_element to resort the
vector to find the "other" middle value is slightly more expensive than calling
max_element() to retrieve the largest value from the left half of the vector.
This is guaranteed to return the true median, but a future optimization could
remove this second O(n) operation and just return one of the two middle values
rather than their average.