Fix a data race.
It's possile for the main thread to queue frames that are never processed,
because worker threads can process a frame, then see that the
"processingComplete" flag is set and exit without checking whether there are
additional frames to process.

To fix this, only allow worker threads to exit if the unprocessed frame queue
is empty.